+++ src/main/java/Facilities/NewYorkCityFacility.java
@@ -16,8 +16,8 @@ import java.util.Map;
 public final class NewYorkCityFacility implements Facility, Comparable<Facility> {
 
     private String location = "New York City, NY";
-    private long ratePerDay = 14;
-    private long costPerDay = 300;
+    private Integer ratePerDay = 14;
+    private Integer costPerDay = 300;
     private ArrayList<FacilityEdge> neighbors = new ArrayList<>();
 
     private HashMap<Item, Integer> activeInventory = new HashMap<>();
@@ -56,7 +56,7 @@ public final class NewYorkCityFacility implements Facility, Comparable<Facility>
     }
 
     @Override
-    public Long getRatePerDay() {
+    public Integer getRatePerDay() {
         return ratePerDay;
     }
 
@@ -66,7 +66,7 @@ public final class NewYorkCityFacility implements Facility, Comparable<Facility>
     }
 
     @Override
-    public Long getCostPerDay() {
+    public Integer getCostPerDay() {
         return costPerDay;
     }
